Moved Property into its own file.
